+++ src/components/cards/svg/astra.jsx
@@ -1,0 +1,31 @@
+export const AstraThemeSvg = () => {
+	return (
+		<svg
+			width="21"
+			height="20"
+			viewBox="0 0 21 20"
+			fill="none"
+			xmlns="http://www.w3.org/2000/svg"
+		>
+			<path
+				fillRule="evenodd"
+				clipRule="evenodd"
+				d="M10.5 20C16.0229 20 20.5 15.5228 20.5 10C20.5 4.47715 16.0229 0 10.5 0C4.97716 0 0.5 4.47715 0.5 10C0.5 15.5228 4.97716 20 10.5 20ZM10.379 4.43553C9.53231 6.22172 8.68559 8.00991 7.83888 9.79805C6.99206 11.5865 6.14519 13.3749 5.29838 15.1613H7.47578C8.1613 13.7783 8.84678 12.3932 9.53225 11.0081C10.2177 9.62302 10.9032 8.23794 11.5887 6.85487L10.379 4.43553ZM11.5887 11.0483C11.9374 10.3225 12.2862 9.59679 12.637 8.87103C13.149 9.91933 13.6591 10.9677 14.1691 12.016C14.6793 13.0644 15.1894 14.1129 15.7015 15.1613H13.3628C13.2298 14.8509 13.0946 14.5424 12.9596 14.2339C12.8245 13.9255 12.6894 13.617 12.5563 13.3065H10.5402H10.4999L10.5402 13.2258C10.8911 12.5 11.2399 11.7742 11.5887 11.0483Z"
+				fill="url(#paint0_linear_9131_38968)"
+			/>
+			<defs>
+				<linearGradient
+					id="paint0_linear_9131_38968"
+					x1="20.5"
+					y1="-5.96046e-07"
+					x2="0.499998"
+					y2="20"
+					gradientUnits="userSpaceOnUse"
+				>
+					<stop stopColor="#B147E1" />
+					<stop offset="1" stopColor="#5236DD" />
+				</linearGradient>
+			</defs>
+		</svg>
+	);
+};
