+++ src/components/cards/svg/presto-player.jsx
@@ -1,0 +1,144 @@
+export const PrestoPlayerSvg = () => {
+	return (
+		<svg
+			width="21"
+			height="20"
+			viewBox="0 0 21 20"
+			fill="none"
+			xmlns="http://www.w3.org/2000/svg"
+		>
+			<path
+				d="M13.6094 9.93105L19.3338 6.60156V13.2606L13.6094 9.93105Z"
+				fill="#3A30E5"
+			/>
+			<path
+				d="M13.6094 3.32812L19.3338 6.65763L13.6094 9.98711V3.32812Z"
+				fill="#3058E5"
+			/>
+			<path
+				d="M13.615 3.32812V9.9287L7.89062 6.59922L13.615 3.32812Z"
+				fill="#4067EE"
+			/>
+			<path
+				d="M13.6697 3.32948L7.94531 6.65899V0L13.6697 3.32948Z"
+				fill="#476EF9"
+			/>
+			<path
+				d="M7.88845 6.66016V13.3191L2.16406 9.98964L7.88845 6.66016Z"
+				fill="#624AF6"
+			/>
+			<path
+				d="M7.88845 0V6.65899L2.16406 3.32948L7.88845 0Z"
+				fill="#4067EE"
+			/>
+			<path
+				d="M7.88845 13.2578V19.9168L2.16406 16.5873L7.88845 13.2578Z"
+				fill="#7C3AED"
+			/>
+			<path
+				d="M13.6094 9.92969L19.3338 13.2592L13.6094 16.5887V9.92969Z"
+				fill="#4F46E5"
+			/>
+			<path
+				d="M2.16406 9.92969L7.88845 13.2592L2.16406 16.5887V9.92969Z"
+				fill="#814AF7"
+			/>
+			<path
+				d="M2.16406 3.32812L7.88845 6.65763L2.16406 9.98711V3.32812Z"
+				fill="#3058E5"
+			/>
+			<path
+				d="M13.615 9.92969V16.5887L7.89062 13.2592L13.615 9.92969Z"
+				fill="#624AF6"
+			/>
+			<path
+				d="M13.6094 9.93105L19.3338 6.60156V13.2606L13.6094 9.93105Z"
+				fill="#3A30E5"
+			/>
+			<path
+				d="M13.6094 3.32812L19.3338 6.65763L13.6094 9.98711V3.32812Z"
+				fill="#3058E5"
+			/>
+			<path
+				d="M13.615 3.32812V9.9287L7.89062 6.59922L13.615 3.32812Z"
+				fill="#4067EE"
+			/>
+			<path
+				d="M13.6697 3.32948L7.94531 6.65899V0L13.6697 3.32948Z"
+				fill="#476EF9"
+			/>
+			<path
+				d="M7.88845 6.66016V13.3191L2.16406 9.98964L7.88845 6.66016Z"
+				fill="#624AF6"
+			/>
+			<path
+				d="M7.88845 0V6.65899L2.16406 3.32948L7.88845 0Z"
+				fill="#4067EE"
+			/>
+			<path
+				d="M7.88845 13.2578V19.9168L2.16406 16.5873L7.88845 13.2578Z"
+				fill="#7C3AED"
+			/>
+			<path
+				d="M13.6094 9.92969L19.3338 13.2592L13.6094 16.5887V9.92969Z"
+				fill="#4F46E5"
+			/>
+			<path
+				d="M2.16406 9.92969L7.88845 13.2592L2.16406 16.5887V9.92969Z"
+				fill="#814AF7"
+			/>
+			<path
+				d="M2.16406 3.32812L7.88845 6.65763L2.16406 9.98711V3.32812Z"
+				fill="#3058E5"
+			/>
+			<path
+				d="M13.615 9.92969V16.5887L7.89062 13.2592L13.615 9.92969Z"
+				fill="#624AF6"
+			/>
+			<path
+				d="M13.6094 9.93105L19.3338 6.60156V13.2606L13.6094 9.93105Z"
+				fill="#3A30E5"
+			/>
+			<path
+				d="M13.6094 3.32812L19.3338 6.65763L13.6094 9.98711V3.32812Z"
+				fill="#3058E5"
+			/>
+			<path
+				d="M13.615 3.32812V9.9287L7.89062 6.59922L13.615 3.32812Z"
+				fill="#4067EE"
+			/>
+			<path
+				d="M13.6697 3.32948L7.94531 6.65899V0L13.6697 3.32948Z"
+				fill="#476EF9"
+			/>
+			<path
+				d="M7.88845 6.66016V13.3191L2.16406 9.98964L7.88845 6.66016Z"
+				fill="#624AF6"
+			/>
+			<path
+				d="M7.88845 0V6.65899L2.16406 3.32948L7.88845 0Z"
+				fill="#4067EE"
+			/>
+			<path
+				d="M7.88845 13.2578V19.9168L2.16406 16.5873L7.88845 13.2578Z"
+				fill="#7C3AED"
+			/>
+			<path
+				d="M13.6094 9.92969L19.3338 13.2592L13.6094 16.5887V9.92969Z"
+				fill="#4F46E5"
+			/>
+			<path
+				d="M2.16406 9.92969L7.88845 13.2592L2.16406 16.5887V9.92969Z"
+				fill="#814AF7"
+			/>
+			<path
+				d="M2.16406 3.32812L7.88845 6.65763L2.16406 9.98711V3.32812Z"
+				fill="#3058E5"
+			/>
+			<path
+				d="M13.615 9.92969V16.5887L7.89062 13.2592L13.615 9.92969Z"
+				fill="#624AF6"
+			/>
+		</svg>
+	);
+};
