+++ src/components/cards/svg/surecart.jsx
@@ -1,0 +1,18 @@
+export const SureCartSvg = () => {
+	return (
+		<svg
+			width="21"
+			height="20"
+			viewBox="0 0 21 20"
+			fill="none"
+			xmlns="http://www.w3.org/2000/svg"
+		>
+			<path
+				fillRule="evenodd"
+				clipRule="evenodd"
+				d="M10.5 20C16.0229 20 20.5 15.5228 20.5 9.99995C20.5 4.47713 16.0229 0 10.5 0C4.97716 0 0.5 4.47713 0.5 9.99995C0.5 15.5228 4.97716 20 10.5 20ZM10.5431 4.99997C9.74015 4.99997 8.62895 5.45918 8.06117 6.02561L6.51908 7.56407H14.2054L16.7755 4.99997H10.5431ZM12.9259 13.9743C12.3582 14.5407 11.247 14.9999 10.444 14.9999H4.21157L6.7817 12.4358H14.468L12.9259 13.9743ZM15.4241 8.84611H5.23681L4.7556 9.32688C3.61617 10.3525 3.95412 11.1538 5.54983 11.1538H15.7647L16.246 10.673C17.3744 9.65344 17.0198 8.84611 15.4241 8.84611Z"
+				fill="#01824C"
+			/>
+		</svg>
+	);
+};
