+++ src/components/radio-button-group/radio-button-group.jsx
@@ -160,6 +160,7 @@ const RadioButtonComponent = (
 	},
 	ref
 ) => {
+	const { buttonWrapperClasses } = props;
 	const providerValue = useRadioButton();
 	const {
 		name,
@@ -267,7 +268,8 @@ const RadioButtonComponent = (
 		<label
 			className={ cn(
 				'inline-flex items-center relative cursor-pointer transition-all duration-300',
-				!! label && 'items-start justify-between min-w-[180px] ',
+				!! label && 'items-start justify-between min-w-[180px]',
+				buttonWrapperClasses,
 				borderOn &&
 					'border border-border-subtle border-solid rounded-md shadow-sm hover:ring-2 hover:ring-border-interactive',
 				borderOn && checkedValue && 'ring-2 ring-border-interactive',
