+++ src/components/radio-button-group/radio-button-group.stories.js
@@ -188,6 +188,14 @@ export default {
 				defaultValue: { summary: 'false' },
 			},
 		},
+		buttonWrapperClasses: {
+			description:
+				'`RadioButton.Button` : Custom classes to be applied to the button wrapper.',
+			control: {
+				type: 'text',
+			},
+			defaultValue: 'bg-white',
+		},
 	},
 };
